Move implementation of transfer to cc file
Use drake::Parallelism
