+++ multibody/mpm/sparse_grid.cc
@@ -169,13 +169,21 @@ Vector3<int> SparseGrid<T>::OffsetToCoordinate(uint64_t offset) const {
  each other in physical space are close to each other in memory. */
 template <typename T>
 void SparseGrid<T>::Sort(std::vector<ParticleIndex>* particles) {
-  pss::parallel_stable_sort(particles->begin(), particles->end(),
-                            [](const ParticleIndex& a, const ParticleIndex& b) {
-                              if (a.base_node_offset == b.base_node_offset) {
-                                return a.index < b.index;
-                              }
-                              return a.base_node_offset < b.base_node_offset;
-                            });
+  // pss::parallel_stable_sort(particles->begin(), particles->end(),
+  //                           [](const ParticleIndex& a, const ParticleIndex&
+  //                           b) {
+  //                             if (a.base_node_offset == b.base_node_offset) {
+  //                               return a.index < b.index;
+  //                             }
+  //                             return a.base_node_offset < b.base_node_offset;
+  //                           });
+  std::sort(particles->begin(), particles->end(),
+            [](const ParticleIndex& a, const ParticleIndex& b) {
+              if (a.base_node_offset == b.base_node_offset) {
+                return a.index < b.index;
+              }
+              return a.base_node_offset < b.base_node_offset;
+            });
 }
 
 template <typename T>
@@ -183,7 +191,9 @@ void SparseGrid<T>::SortParticleIndices(ParticleData<T>* data) {
   const std::vector<Vector3<T>>& particle_positions = data->x;
   const int num_particles = particle_positions.size();
   particles_.resize(num_particles);
+#if defined (_OPENMP)
 #pragma omp parallel for
+#endif
   for (int p = 0; p < num_particles; ++p) {
     const Vector3<int> base_node =
         mpm::internal::base_node<T>(particle_positions[p] / dx_);
